Update libtool.m4
This updates to a more recent libtool version 2.4.6 with fixes for many
modern OSes. For example, FreeBSD 12 would not compile dynamic libraries
due to an old freebsd1* that detected FreeBSD 12 as being FreeBSD 1.

The Ubuntu version of `libtoolize -c` was used, which is slightly
different from the official GNU libtool.